feat: move icons new file
