+++ app/page.tsx
@@ -1,7 +1,12 @@
+import { GlobalLayout } from "~/components/global-layout.components";
+
+
 export default function Home() {
   return (
     <div>
-     <p>Home Page</p>
+      <GlobalLayout>
+        <p>Home Page</p>
+      </GlobalLayout>
     </div>
   )
 }
