+++ app/wallet/layout.tsx
@@ -1,0 +1,16 @@
+import { GlobalLayout } from "~/components/global-layout.components"
+
+export default function WalletLayout({
+    children,
+  }: {
+    children: React.ReactNode
+  }) {
+    return (
+      <div>     
+        <GlobalLayout>
+          {children}
+        </GlobalLayout>
+      </div>
+    )
+  }
+
