+++ components/crypto-select/crypto-select.type.ts
