+++ components/header/header.component.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { Fragment, useState } from 'react'
 import { Disclosure, Menu, Transition } from '@headlessui/react'
-import { Blockmatic, BlockmaticIcon, MenuIcons, BellIcon, CaretRight, CaretDown } from '~/ui/icons'
-import XcloseIcon from '~/ui/icons/XcloseIcon'
+import { Blockmatic, BlockmaticIcon, MenuIcons, BellIcon, CaretRight, CaretDown } from '~/icons'
+import XcloseIcon from '~/icons/XcloseIcon'
 import Image from 'next/image'
 import { Sidebar } from '../sidebar/sidebar.component'
 import SearchBar from '../search-bar/search-bar.component'
