+++ components/search-bar/search-bar.component.tsx
@@ -1,7 +1,7 @@
 import { Fragment, useState } from 'react';
 import { Listbox, Transition } from '@headlessui/react';
 import { CheckIcon } from '@heroicons/react/20/solid';
-import { CaretDown, CaretRight, MagnifyingGlassIcon } from '~/ui/icons';
+import { CaretDown, CaretRight, MagnifyingGlassIcon } from '~/icons';
 
 
 const coins = [
