+++ icons/MagnifyingGlassIcon.tsx
